+++ libft/ft_strlen.c
@@ -1,13 +1,15 @@
-unsigned long ft_strlen(const char *s)
+#include "libft.h"
+
+size_t ft_strlen(const char *s)
 {
-    unsigned long length;
+    size_t length;
 
     length = 0;
     while (s[length])
         length++;
     return (length);
 }
-
+/*
 #include <string.h>
 #include <stdio.h>
 
@@ -16,3 +18,4 @@ int main()
     printf("Original: %d\n", strlen("Hello World!"));
     printf("Mine: %d\n", ft_strlen("Hello World!"));
 }
+*/
